Add 'assigns' field to %EventPipe.Event{}
The ShopifyAPI library itself does not make use of this field, making it
a useful mechanism to pass through additional contextual information
when using the EventPipe to perform actions asynchronously.

As an example, it can be used to make sure application-specific
information is available when an event's callback function is invoked.