Add method, duration to API request logging
This aims to improve the usefulness of the debug logs provided for REST
requests against the Shopify API, by including both the request method
and the duration of the request. The latter is achieved using Erlang's
`timer:tc/`, which returns the duration of the provided function in μs.

The leading `Elixir.` is also stripped from the logged module name.

Before:

    [debug] Elixir.ShopifyAPI.REST.Request requesting "https://shop-name.myshopify.com/admin/api/2019-04/locations.json"

After:

    [debug] ShopifyAPI.REST.Request GET https://shop-name.myshopify.com/admin/api/2019-04/locations.json [627ms]