Remove email address from the Customer Authenticator token
This change updates the documentation for the ShopifyAPI.Plugs.CustomerAuthenticator to remove the email address from the JSON token. This key is not being used and since it constitutes PII, should ideally not be included in the payload. The `id` element is already part of the payload and can serve as a unique customer identifier instead.