Return custom errors in ShopifyAPI.Bulk.Query
We want to be able to catch specific errors raised by 
Bulk.process_stream!/3 and Bulk.process!/3 in their callers. 

Bulk.Query.exec/3 previously passed errors down to be raised later in
the chain - Bulk.Query.exec!/3 now raises its own errors.

The :no_objects condition needed to be moved to the bottom to satisfy
dialyzer - dialyzer also had big problems with using “<>” in a match,
which is why it lives in a function instead of in exec!/3’s else block