Update To Handle Route With App Name In Path
This update allows the routing logic to handle app names either provided
as the last part of the path component of a URI or -- as it currently
does -- via the 'app' parameter. The modification doesn't add any new
functionality, but takes advantage of existing functionality & makes it
available in a different scope.