Get OAuth Flow Working
This modification primarily attempts to make working with the OAuth flow
easier for an integrator. It also fixes the OAuth flow so that there's a
full round trip for both installation and app authorization.

I believe before we were never doing the full round trip for
standard authorization.

Additionally, we've added a couple of hooks that use the ShopifyAPI
module configuration pattern to allow integrators to hook into the auth
flow without needing to understand the nitty-gritty details of how it
works.

But because there is always going to be some amount of manual
configuration required for the app to be installed in the Shopify store
and that necessitates some underlying understanding of how the OAuth
flow works, we've attempted to add some documentation both in the README
as well as in the routes file to help integrators with understanding how
things work.

The modifications in the routes file break what was there before -- the
routes that were defined are no longer valid. I believe that this will
not really be a dealbreaker for most, as I believe that the former
implementation required integrators to have their own self-baked
solution in place.

This modification hopefully makes it easier to integrate the library
into an application authorization flow. It includes addition of callback
hooks dedicated to making authorization more simple and more
explicit.