Add basic request id logging
One of elixir-wholesale's bulk operations is timing out often (but not always) and Shopify support requires the request id of request to create the bulk operation before investigating. This commit is meant only to be temporarily deployed - if we want to get this logging into the app, we'll need to label these logs somehow with the query being called.